Moved bitarray.py logic into WorldSlice.py
